requirements.txt
flask
flask-cors
tensorflow
pillow
numpy
# pip install mysql-connector-python flask